Move DashboardShelf component to its own file
