+++ src/components/dashboard.tsx
@@ -2,8 +2,7 @@ import { HeadingLevels } from "@/types/text";
 import { Heading } from "./ui/heading";
 import { HTMLAttributes } from "react";
 import { Section } from "./ui/section";
-import { VariantProps, cva } from "class-variance-authority";
-import { cn } from "../../lib/utils";
+import { DashboardShelf } from "./dashboard-shelf";
 
 interface DashboardRootLayoutProps {
   children: React.ReactNode;
@@ -46,34 +45,6 @@ interface DashboardHeadingProps extends HTMLAttributes<HTMLHeadingElement> {}
 
 const DashboardHeading = ({ children }: DashboardHeadingProps) => {
   return <Heading level={HeadingLevels.h3}>{children}</Heading>;
-};
-
-const dashboardShelfVariants = cva(
-  "lg:inline-flex lg:w-full lg:flex-row lg:gap-12 lg:gap-8 lg:overflow-x-auto lg:pb-8",
-  {
-    variants: {
-      mobileBehavior: {
-        default:
-          "inline-flex w-full flex-row gap-8 overflow-x-auto pb-8 lg:gap-12",
-        "one-col": "grid grid-cols-1 gap-12 lg:gap-16",
-        "two-col": "grid grid-cols-2 gap-x-12 gap-y-8",
-        "three-col": "grid grid-cols-3 gap-x-12 gap-y-8",
-      },
-    },
-    defaultVariants: { mobileBehavior: "default" },
-  }
-);
-
-interface DashboardShelfProps
-  extends HTMLAttributes<HTMLElement>,
-    VariantProps<typeof dashboardShelfVariants> {}
-
-const DashboardShelf = ({ mobileBehavior, children }: DashboardShelfProps) => {
-  return (
-    <section className={cn(dashboardShelfVariants({ mobileBehavior }))}>
-      {children}
-    </section>
-  );
 };
 
 interface DashboardContentShelfProps {
